Nuke interface and factory
