+++ library/src/main/java/com/novoda/downloadmanager/NotificationCreator.java
@@ -1,23 +1,50 @@
 package com.novoda.downloadmanager;
 
+import android.app.Notification;
 import android.app.NotificationChannel;
 import android.content.Context;
 import android.os.Build;
 import android.support.annotation.RequiresApi;
+import android.support.v4.app.NotificationCompat;
 
-public interface NotificationCreator<T> {
-    NotificationInformation createNotification(T notificationPayload);
+public class NotificationCreator<T> {
+
+    private final Context applicationContext;
+    private final String channelId;
+    private final String userFacingChannelDescription;
+    private final NotificationCustomizer<T> notificationCustomizer;
+    @Importance
+    private final int importance;
+
+    public NotificationCreator(Context context,
+                               String channelId,
+                               String userFacingChannelDescription,
+                               NotificationCustomizer<T> customizer,
+                               @Importance int importance) {
+        this.applicationContext = context.getApplicationContext();
+        this.channelId = channelId;
+        this.userFacingChannelDescription = userFacingChannelDescription;
+        this.notificationCustomizer = customizer;
+        this.importance = importance;
+    }
+
+    NotificationInformation createNotification(final T notificationPayload) {
+        return new NotificationInformation() {
+            @Override
+            public int getId() {
+                return notificationPayload.hashCode();
+            }
+
+            @Override
+            public Notification getNotification() {
+                NotificationCompat.Builder builder = new NotificationCompat.Builder(applicationContext, channelId);
+                return notificationCustomizer.customNotificationFrom(builder, notificationPayload);
+            }
+        };
+    }
 
     @RequiresApi(Build.VERSION_CODES.O)
-    NotificationChannel createNotificationChannel();
-
-    class Factory {
-        private Factory() {
-            // Uses static methods
-        }
-
-        public static <T> NotificationCreator<T> build(Context context, String channelId, String userFacingChannelDescription, NotificationCustomizer<T> customizer, @Importance int importance) {
-            return new LiteNotificationCreator<>(context, channelId, userFacingChannelDescription, customizer, importance);
-        }
+    NotificationChannel createNotificationChannel() {
+        return new NotificationChannel(channelId, userFacingChannelDescription, importance);
     }
 }
